rename: change package name from moreinfo to more_info
