+++ thismuch/src/main/java/com/project/thismuch/more_info/MoreInfoService.java
@@ -1,4 +1,4 @@
-package com.project.thismuch.moreinfo;
+package com.project.thismuch.more_info;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
